requirements.txt
nltk
pytest-mock